requirements.txt
streamlit==1.37.0
validators==0.28.3
requests==2.32.3
Pillow==10.4.0
pyzbar==0.1.9